test: OperationTests 내부 FormulaTests클래스를 별도 파일로 분리
