+++ Calculator/CalculatorTests/OperationTests.swift
@@ -86,46 +86,3 @@ class OperatorTests: XCTestCase {
         XCTAssertEqual(finalCalculatedAnswer, 3.7 / 4.3 * 11.8)
     }
 }
-
-class FormulaTests: XCTestCase {
-    var formula: Formula = Formula()
-    
-    override func setUp() {
-        formula.operands.enqueue(3.7)
-        formula.operands.enqueue(4.9)
-        formula.operands.enqueue(0.0)
-        formula.operands.enqueue(12.4)
-        formula.operands.enqueue(-342.6)
-    }
-    
-    override func tearDown() {
-        formula = Formula()
-    }
-    
-    func testCalculateResult() throws {
-        formula.operators.enqueue(.add)
-        formula.operators.enqueue(.subtract)
-        formula.operators.enqueue(.devide)
-        formula.operators.enqueue(.multiply)
-        
-        do {
-            let result = try formula.result()
-            XCTAssertEqual(result, ((3.7 + 4.9) - 0.0) / 12.4 * -342.6)
-        } catch QueueError.isEmpty {
-            print(QueueError.isEmpty.localizedDescription)
-        } catch OperationError.devidedByZero {
-            print(OperationError.devidedByZero)
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
-    
-    func testCalculateResultWhereDevideByZero() throws {
-        formula.operators.enqueue(.add)
-        formula.operators.enqueue(.devide)
-        formula.operators.enqueue(.subtract)
-        formula.operators.enqueue(.multiply)
-        
-        XCTAssertThrowsError(try formula.result())
-    }
-}
